requirements.txt
selenium==4.9.1
pytest==7.3.1
allure_pytest==2.13.2
openpyxl==3.1.2
pytesseract==0.3.10
Pillow==9.5.0
requests==2.30.0
mysql-connector==2.2.9
cx-Oracle==8.3.0